requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-import-export==3.3.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
gast==0.5.4
geoip2==4.7.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
iso8601==2.1.0
jdatetime==4.1.1
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.3
maxminddb==2.4.0
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
sequential==1.0.0
signals==0.0.2
six==1.16.0
sklearn==0.0.post11
sqlparse==0.4.4
tablib==3.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
